refactor: move the engine's own strings out of the screens package
ScreenText, the keys of every string this engine shows, lived among the
widgets and was read from three places outside them, the host report,
the chain's compiling label and Sodium's config entry, so the frame and
the root of the mod both named the screens package for a list of words.
The list goes to the root package beside the host report; the widgets
import it from there. Nothing in the list changes.
