+++ common/src/main/java/dev/vitrail/screen/LinkWidget.java
@@ -1,5 +1,7 @@
 package dev.vitrail.screen;
 
+import dev.vitrail.ScreenText;
+
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
