+++ common/src/main/java/dev/vitrail/screen/SettingsScreen.java
@@ -1,14 +1,15 @@
 package dev.vitrail.screen;
 
-import dev.vitrail.Vitrail;
+import dev.vitrail.pack.load.PackLoader;
 import dev.vitrail.pack.menu.MenuValues;
 import dev.vitrail.pack.source.PackLang;
-import dev.vitrail.pack.load.PackLoader;
 import dev.vitrail.render.PackChain;
+import dev.vitrail.ScreenText;
 import dev.vitrail.settings.PackFile;
 import dev.vitrail.settings.PackSession;
 import dev.vitrail.settings.SettingsFile;
 import dev.vitrail.uniform.Smoothed;
+import dev.vitrail.Vitrail;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
