+++ common/src/main/java/dev/vitrail/sodium/ConfigEntry.java
@@ -4,8 +4,8 @@ import dev.vitrail.render.ModuleCache;
 import dev.vitrail.render.PackChain;
 import dev.vitrail.render.StartupGuard;
 import dev.vitrail.render.TerrainDraw;
-import dev.vitrail.screen.ScreenText;
 import dev.vitrail.screen.SettingsScreen;
+import dev.vitrail.ScreenText;
 import dev.vitrail.settings.GraphicsApiChoice;
 import dev.vitrail.settings.PackFile;
 import dev.vitrail.Vitrail;
